First pass at mwlib, just checking in so others can hack on it.
Append ?parser=2 to the URL to get the old functionality.

Not working: Link URLs, Templates (not in the DB?), Images, Colored links, some rendering errors like REDIRECTs scattered through the code.

But, the overall rendering seems a little better (no list indent problems).  Need to fix bugs and test the heck out of this for Tuesday!