Resolved server/browse startup race
In an attempt to solve a race between the browse activity and the wikibrowser server startup Daf blocked the activity thread until server.py printed a line to its stdout indicating its state.  This did not solve the race, because the print statement had to come before server startup.  Now we spawn the server in a new thread and indicate readiness after the server thread has been run, resolving the race.