Implement a custom setup.py
The help activity is different to other activities, we store in git
the source .rst files, but no the generated html.
Also, we have the images in the "images" directory, but a symlink
is created inside the html directory.
This custom setup.py takes care of include only the needed files.