Add all the html files translated to Spanish
The translated files are generated with the .rst and the .po for the language.
Both files need match, if the rst file has changed, and the .po was
not updated the html will be generated with a broken translation.
The solution is add the translated html files to git,
and is different to the case of the English html files, that are not in git,
and are generated at activity build time.